requirements.txt
flask_restful
PyMySQL
flask-bcrypt
flask-jwt-extended
flask-cors
google-cloud-storage==1.31.0
gunicorn==20.0.4